Various Cleanup
  * Organize the interface to follow LibSodium's docs.
    - No code changes, just moved declarations around and added comments to
assist as the number of declarations begin to grow.

  * Standardize NaCl.Sodium formatting.
    - Formatting was all over the place in NaCl.Sodium, this attempts to
standarize the formatting such that the right margin is 80 characters
and long argment lists start on the next line after the method name.

  * Remove redundant modifiers from NaCl.Sodium.
    - All methods and fields in java interfaces are default 'public'.
    - All fields in java interfaces are 'static final'

  * Make the static init() function synchronized.

  * Change array NaCl.Sodium len values to ints.
    - Java arrays are always indexed by ints and thus limited to Integer.MAX
elements, in theory. I was worried that the array length values were
defined as longs for a performance reason with jnr-ffi but in testing I
couldn't see any difference.
    It is better to define them as ints here to be consistent with the Java
language, I think.

  * Add missing @In annotations to NaCl.Sodium.

  * Use org.junit.Assert instead of junit.framework.Assert.