Add deriveKey method for Password
This returns a byte array rather than a string like the existing hash
functions. This is important when using pwhash to derive a key as
opposed to password storage, because you will want to clear the key from
memory as soon as possible after deriving and using it. If the derived
key is converted to an immutable string, this is not possible.

The old key derivation methods have been left in place; hopefully the
name difference will prompt users to choose the byte array methods for
key derivation and the string methods for password storage.